* more cleanup * more fixes in BitmapBlock::FindNextMarked() and BitmapBlock::FindPreviousMarked()
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@39242 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/file_systems/ext2/Inode.cpp
@@ -938,7 +938,7 @@ Inode::_ShrinkDataStream(Transaction& transaction, off_t size)
 	if (size > minSize) {
 		// No need to allocate more blocks
 		TRACE("Inode::_ShrinkDataStream(): No need to allocate more blocks\n");
-		TRACE("Inode::_ShrinkDataStream(): Setting size to %ld\n", (long)size);
+		TRACE("Inode::_ShrinkDataStream(): Setting size to %lld\n", size);
 		fNode.SetSize(size);
 		return B_OK;
 	}
